django_examples/hello_django/pyproject.toml
[project]
name = "hello-django"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=6.0.1",
    "django-crispy-forms>=2.5",
    "django-extensions>=4.1",
    "factory-boy>=3.3.3",
    "faker>=40.1.2",
    "ipython>=9.10.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
]

